Task-1
Created a program that takes a string as input and extracts everything from it
numbers in a separate array, after which the program prints a string without numbers. and
an array of numbers. Modifies this line so that each word in it,
began and ended with a capital letter. Will find the maximum
values ​​in an array of numbers, and all other numbers to the power of them
index, and write to another array.